feat: Query Planning - Rough First Plumbing
Introduce the concept of a query plan and a query planner that takes a
Select and builds it into a plan.

The MemoryStore now uses the query plan to execute queries.

This is just a very rough, primitive implementation.
The only query optimization present is to optimize `WHERE id = XXX`,
which was already done in the MemoryStore.

Next step is to implement various query optimizations, including using
indexes.