Introduce QueryPlan + Optimizer infrastructure
QueryOp was a weird mix between a tree and a sequential list of
operations.

The query plan must really be a tree, not a list.

Hence the QueryOp is refactored into a tree.

A new Optimizer trait is added for defining optimization rules.

A single rule is present for now that implements the previous
optimization for single entity selects.

The memory backend query logic is also updated.

NOTE: as a side effect, the memory backend can now properly handle index
selects  and scans.